+++ crates/jstz_api/src/lib.rs
@@ -1,13 +1,14 @@
 mod console;
-mod kv;
-
 pub mod encoding;
 pub mod http;
 pub mod idl;
+pub mod js_log;
+mod kv;
 pub mod random;
 pub mod url;
 pub mod urlpattern;
-pub use console::{set_js_logger, ConsoleApi, JsLog, LogData, LogLevel};
+
+pub use console::ConsoleApi;
 pub use kv::Kv;
 pub use kv::KvApi;
 pub use kv::KvValue;
