+++ crates/jstz_proto/src/lib.rs
@@ -3,6 +3,7 @@ mod error;
 
 pub mod context;
 pub mod executor;
+pub mod js_logger;
 pub mod operation;
 pub mod receipt;
